Updated kernel/ring_buffer.c wakeups to use schedule_work()
There were cases where the tracing framework, particularly trace_pipe
would block forever upon read even though there were events available.
This happened when the buffer became empty and the reader was put into a wait
queue but was never woken up when there were events available. This in turn
occurs if the system was in NO_HZ state and the timer interrupt which is
responsible for running these irq workloads was never triggered.

This commit replaces the IRQ based work scheduling with schedule_work and
schedule_work_on. This fixes the problem as these tasks are added to the
global/CPU runqueues respectively and are scheduled eventually.